// src/store/main/system/type.ts
export interface IsystemState {
  //  用户管理
  usersList: any
  usersCount: number
  // 部门管理
  departmentTotalCount: number
  departmentList: any
  // 菜单管理
  menuList: any
  menuCount: number
  // 角色管理
  roleList: any
  roleCount: number
  // 商品分类
  categoryTotalCount: number
  categoryList: any[]
  // 商品信息
  goodsList: any
  goodsCount: number
  // 故事列表
  storyList: any
  storyTotalCount: number
}